+++ CHANGELOG.md
@@ -2,6 +2,40 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [7.0.2](https://github.com/jsverse/transloco-keys-manager/compare/v7.0.1...v7.0.2) (2025-07-20)
+
+
+### Bug Fixes
+
+* 🐛 angular-utils breaking TKM ([e82bdb6](https://github.com/jsverse/transloco-keys-manager/commit/e82bdb6b65eeda86c075243abaa0bd5119a4e108)), closes [#231](https://github.com/jsverse/transloco-keys-manager/issues/231)
+
+### [7.0.1](https://github.com/jsverse/transloco-keys-manager/compare/v7.0.0...v7.0.1) (2025-07-18)
+
+## [7.0.0](https://github.com/jsverse/transloco-keys-manager/compare/v6.2.2...v7.0.0) (2025-06-23)
+
+
+### ⚠ BREAKING CHANGES
+
+* 🧨 Minimum Angular version is 20 + Minimum node version is 20.19.0
+
+This reverts commit fb965a6dc6e1429f3991e2a7bf53d55801821c47.
+
+### Features
+
+* 🎸 Support angular 20 ([#225](https://github.com/jsverse/transloco-keys-manager/issues/225)) ([26162b9](https://github.com/jsverse/transloco-keys-manager/commit/26162b91d17bbf35cdb6b220f2d482aba5fc3786))
+
+### [6.2.2](https://github.com/jsverse/transloco-keys-manager/compare/v6.2.1...v6.2.2) (2025-06-23)
+
+### [6.2.1](https://github.com/jsverse/transloco-keys-manager/compare/v6.2.0...v6.2.1) (2025-06-23)
+
+## [6.2.0](https://github.com/jsverse/transloco-keys-manager/compare/v6.1.0...v6.2.0) (2025-06-19)
+
+
+### Features
+
+* 🎸 Support angular 20 ([#225](https://github.com/jsverse/transloco-keys-manager/issues/225)) ([6c121df](https://github.com/jsverse/transloco-keys-manager/commit/6c121df159ffa32bcec7e8987d58dd495da8516b))
+* 🎸 support signal-based API ([#223](https://github.com/jsverse/transloco-keys-manager/issues/223)) ([a0d75c3](https://github.com/jsverse/transloco-keys-manager/commit/a0d75c3c32fabee5a23ea05d274dbc9ffe10de1f))
+
 ## [6.1.0](https://github.com/jsverse/transloco-keys-manager/compare/v6.0.0...v6.1.0) (2024-12-19)